tests/requirements-test.txt
pytest>=8.0
requests>=2.31
paho-mqtt>=1.6
pymongo>=4.7